Rename FileUpload file, start adapting for legacy Uploader
